More tweaks to the error handler
