+++ cogs/events/error_handler.py
@@ -1,0 +1,101 @@
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING
+
+import discord
+from discord.ext import commands
+
+from utils import BlacklistedGuild, BlacklistedUser, Embed, FeatureDisabled, LagContext, UnderMaintenance, better_string
+
+if TYPE_CHECKING:
+    from bot import Lagrange
+log: logging.Logger = logging.getLogger(__name__)
+
+
+class Errors(commands.Cog):
+    def __init__(self, bot: Lagrange) -> None:
+        self.bot = bot
+
+    def clean_error_permission(self, permissions: list[str], *, seperator: str, prefix: str) -> str:
+        return better_string((prefix + (perm.replace('_', ' ')).capitalize() for perm in permissions), seperator=seperator)
+
+    @commands.Cog.listener()
+    async def on_command_error(self, ctx: LagContext, error: commands.CommandError) -> None | discord.Message:
+        if isinstance(error, BlacklistedUser | BlacklistedGuild | UnderMaintenance | FeatureDisabled):
+            if (isinstance(ctx.channel, discord.DMChannel) and isinstance(error, BlacklistedUser)) or isinstance(
+                error,
+                UnderMaintenance | FeatureDisabled,
+            ):
+                await ctx.reply(content=str(error))
+            elif ctx.guild and isinstance(error, BlacklistedGuild):
+                await ctx.guild.leave()
+            return None
+
+        error_messages = {
+            commands.NoPrivateMessage: 'This command cannot be used in DMs',
+            commands.NotOwner: 'You cannot run this command. This command is reserved for the developers of the bot.',
+            commands.UserInputError: 'The input you have provided is invalid.',
+            commands.NSFWChannelRequired: 'This command can only be used in NSFW channels.',
+            commands.PrivateMessageOnly: 'This command can only be used in DMs.',
+            commands.BadArgument: str(error),
+            commands.TooManyArguments: str(error),
+        }
+
+        if (
+            not ctx.command
+            or hasattr(ctx.command, 'on_error')
+            or (ctx.cog and ctx.cog._get_overridden_method(ctx.cog.cog_command_error))
+        ):
+            return None
+
+        elif isinstance(error, commands.CommandOnCooldown):
+            return await ctx.send(
+                f'Command on cooldown! You can use this command after {error.retry_after:.2f} seconds.',
+                delete_after=error.retry_after,
+            )
+
+        elif isinstance(error, commands.MissingRequiredArgument):
+            content = better_string(
+                (
+                    f'You did not provide a {error.param.name} argument.',
+                    (f'-# - {self.bot.help_command.get_command_signature(ctx.command)}' if self.bot.help_command else None),
+                ),
+                seperator='\n',
+            )
+            embed = Embed(
+                ctx=ctx,
+                title='<:redTick:1237048136527249510> | Missing argument',
+                description=content,
+                colour=0xFF0000,
+            )
+            return await ctx.send(embed=embed)
+        elif isinstance(error, commands.MissingPermissions | commands.BotMissingPermissions):
+            content = better_string(
+                (
+                    f'{"You" if isinstance(error, commands.MissingPermissions) else "I"} are missing the following permissions to run this command:',
+                    self.clean_error_permission(list(error.missing_permissions), seperator='\n', prefix='- '),
+                ),
+                seperator='\n',
+            )
+            embed = Embed(
+                ctx=ctx,
+                title='<:redTick:1237048136527249510> | Missing Permissions',
+                description=content,
+                colour=0xFF0000,
+            )
+            return await ctx.reply(embed=embed)
+        elif isinstance(error, tuple(error_messages.keys())):
+            return await ctx.send(
+                embed=Embed(
+                    ctx=ctx,
+                    title='<:status_dnd:1237048172174643200> | Cannot run command',
+                    description=f'{error_messages[type(error)]}',
+                    colour=0xFF0000,
+                ),
+            )
+        return await ctx.reply(f'Placeholder: {error}')
+
+
+async def setup(bot: Lagrange) -> None:
+    await bot.add_cog(Errors(bot))
